Guarded fetch function. Trigger re-authenticate on fetch failure.
