+++ src/routes/Topbar.svelte
@@ -1,8 +1,6 @@
 <script>
 	import { page } from '$app/stores';
 	import { base } from '$app/paths';
-	import logo from '$lib/images/svelte-logo.svg';
-	import github from '$lib/images/github.svg';
 
 	import AuthButton from './AuthButton.svelte';
 	import Profile from './Profile.svelte';
